refactor: reduce redundancy in MovieDTOs
