+++ src/main/java/io/github/janmalch/kino/api/model/movie/NewMovieDto.java
@@ -1,28 +1,8 @@
 package io.github.janmalch.kino.api.model.movie;
 
-public class NewMovieDto {
+public class NewMovieDto extends AbstractMovieDto {
 
-  private String name;
   private long priceCategoryId;
-  private float duration;
-  private int ageRating;
-  private String imageURL;
-
-  public String getImageURL() {
-    return imageURL;
-  }
-
-  public void setImageURL(String imageURL) {
-    this.imageURL = imageURL;
-  }
-
-  public String getName() {
-    return name;
-  }
-
-  public void setName(String name) {
-    this.name = name;
-  }
 
   public long getPriceCategoryId() {
     return priceCategoryId;
@@ -30,22 +10,6 @@ public class NewMovieDto {
 
   public void setPriceCategoryId(long priceCategoryId) {
     this.priceCategoryId = priceCategoryId;
-  }
-
-  public float getDuration() {
-    return duration;
-  }
-
-  public void setDuration(float duration) {
-    this.duration = duration;
-  }
-
-  public int getAgeRating() {
-    return ageRating;
-  }
-
-  public void setAgeRating(int ageRating) {
-    this.ageRating = ageRating;
   }
 
   @Override
